Separating functions test into dedicated file
